fix: move notification icon to core, address review nits
