+++ core/ui/notifications/NotificationBubbleIcon.tsx
